Added Generic Decoder maker.
There is now a function for using generic-sop to construct a `Decoder` using
some approximations of various components. It isn't the recommended way but it
may come in handy.

Replaced the criterion dependency with gauge, as there isn't much cause for the
full power of criterion and gauge prevents this package from having to depend on
the library it aims to surpass. Which is nice.

Added some decoders for pure json structures and maybe, and either data types.

More tests for the generic components is required.