Rework Encoder types to be more in line with Decoder.
The Encoder types were a bit of a mess and were sort of the opposite of the
Decoder type, in name only. So this is largely a cosmetic change with names
changed to protect the applicative...

Some encoder functions have been removed/cleaned up. New encoder functions have
been added to allow for JSON objects with duplicate keys.

Fixed up a bug in the generic newtype encoder/decoder.

Updated the generics to handle the shift in the encoders.

Added some more tests.