Moved benchmarks to their own package.
Benchmarks require a package that won't build on GHC < 8, this isn't acceptable
for Waargonaut so it was moved into its own package.

Added encoder/decoder for scientific number type.

Rewrote Generic deocder functions to use the succinct decoder.

Added tests for new scientific prism, added benchmarks for traversal, manual,
and generic decoders.

Upgraded the version of hw-json to the latest 0.9.