Accidently overhauled object encoding, soz.
Created an explicit encoder for creating objects that says in the type that it
creates objects and that allows for safer modifications.

More importantly it allows for Divisible/Decidable instances to appear.

Much more documentation and tests are now required to update this as the
existing object creation scaffolding should be rethought.