"Fix" the parser round-trip test.
There is a problem with the builder that causes extra escaped numeric values to
be added as characters to the output.

Encoding the character: `\\256` will be built as `\\194\\128`. Which breaks the
round trip rule. This has not been detected because using `Text` as the
input/output type for the round trip hides the error due to the `Text` type
removing spurious non-utf8 valid values.

Additional tests have been added to help identify this problem and for the time
being they are turned off.

Related issue on bytestring: https://github.com/haskell/bytestring/issues/170