Swap Encoder `Builder` for Text over ByteString (#37)
* Swap to lazy Text builder for encoding.

Fix bug in handling of utf8 values. Add more tests. Remove some instances for
`AsJString` that had incorrect Prisms.

* Update doctests for text builder

* Add semigroup constraint, rename test

* Fix typo from merge

* Update version number due to probably breaking change as I removed instances that had functions that might have been in use. It was also a pretty nasty bug.

* Add `parseWith` helper function.

This function makes it a bit easier to piece the bits together when using the
decoder functions.

Add some newlines to the output of the pretty printed cursorhistory.

Add a helper function for people to round trip their encoders and decoders to
make testing a bit easier for others.

* Remove redundant char exclusion from JChar prism.

* Add missing semigroup import.