Expose type inspection functions.
Fix the order that the `Decoder.either` function will attempt the given
decoders, previously it would try the `Left` decoder first.

List, NonEmpty, and 'objectAsKeyValues' decoders now require that they are
positioned at an array or object before attempting to decode. This means that
the list decoder will now only return an empty list when it is looking at an
empty list.