Clean up nix derivation setup.
Pinned the nixpkgs to a head ref of nixos-18.09.
Purged all explicit overrides, save for `hw-json` and `hw-parser`.
Added `dontBenchmark` to `hw-json` to reduce hit on build time.
Removed waargbench, to be uploaded as separate package and cleaned up to use
criterion (must submit PR to remove need for `aeson` dependency, somehow, does
it matter?).