Have a crack at breaking out the `Builder`.
The `Builder` steps are now in their own structure. This is part of a process
that abstracts the builders that are in use, so you can build to `Text` or
`ByteString` and it should make no difference.

There is a property test that given some `Json` both builders will produce (I
think) byte for byte matching output.