Cursor Package Change
The hw-json package now has a dependency on aeson, John has kindly
separated out the cursor component into its own package so we can
benefit from the awesome without incurring a dependency on aeson.

This has resulted in a few changes that include the removal of a type
alias and some other minor changes. I've bumped the major version as
this is a breaking change because a type has gone away. But the by an
large the changes for users should be completely transparent, hopefully.

Cleaned up some of the nix as well to not have so many little files and
direct github links.