Add gObjEncoder, fix some encoder bugs (#70)
* Add gObjEncoder, fix some encoder bugs

Add `onObj'` which is just `onObj` but specialised to `Identity`.

Add `gObjEncoder` for deriving `ObjEncoder` structures for record types only. Using the
`IsRecord` constraint from `record-sop` package. This makes it easier to leverage the
Contravariant functionality of the `ObjEncoder` without losing the benefits of deriving
more trivial encoders.

Added the `FieldNameAsKey` option to the newtype options for generic derived enc/decoders.

Fixes #69 by removing duplicate call to `_optionsFieldName` function. Added regression
test.

Improved the handling of newtype options for generic deriving to give a bit more
flexibility and avoid strangeness wrt to some combinations of options.

Change the building of escaped whitespace chars to actually use the
`escapedWhitespaceChar` function, instead of incorrectly generating an unescaped
character.

* Bump version due to fixes and new functionality

* Add some haddock to gObjEncoder function

* Fixes from feedback on PR.

Correctly bump version to 0.8.0.0 as this is a breaking change because of new constructors
on an exported sum type.

Remove commented code.

Reformat to adhere to own style.

Add a better failure to an "impossible" error case.

* Regenerate nix after cabal file changes

* Add note to haddock for gObjEncoder to mention the overriding of newtype naming options